+++ source/graphics/2d/Textures/Char/CharSet.h
@@ -106,7 +106,7 @@ class CharSet : ListenerObject
 	// Charset spec
 	CharSetSpec* charSetSpec;
 	// Array spec of the charSet
-	uint16 tilesDisplacement;
+	uint32 tilesDisplacement;
 	// Memory displacement
 	uint16 offset;
 	// How many textures are using me
@@ -125,7 +125,7 @@ class CharSet : ListenerObject
 	uint16 getNumberOfChars();
 	void write();
 	void rewrite();
-	void setTilesDisplacement(uint16 tilesDisplacement);
+	void setTilesDisplacement(uint32 tilesDisplacement);
 	void putChar(uint32 charToReplace, uint32* newChar);
 	void putPixel(uint32 charToReplace, Pixel* charSetPixel, BYTE newPixelColor);
 	bool setFrame(uint16 frame);
